refactor(Image): use Modal component in Image
새로 만든 컴포넌트 적용
