+++ src/components/index.ts
@@ -1,3 +1,7 @@
 export { ChatContainer } from './ChatContainer/ChatContainer';
 export { Balloon } from './ChatContainer/Balloon/Balloon';
 export { InputBox } from './InputBox/InputBox';
+export { Image } from './ChatContainer/Balloon/Image/Image';
+export { Button } from './ChatContainer/Balloon/Button/Button';
+export { Text } from './ChatContainer/Balloon/Text/Text';
+export { Youtube } from './ChatContainer/Balloon/Youtube/Youtube';
